backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# HTTP and web scraping
httpx==0.25.2
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.2

# AI and ML
openai==1.3.7
anthropic==0.7.8
stability-sdk==0.8.4
elevenlabs==0.2.26
pillow==10.1.0
numpy==1.25.2

# Video processing
ffmpeg-python==0.2.0
moviepy==1.0.3

# Text processing
jieba==0.42.1
mecab-python3==1.0.6
langdetect==1.0.9
nltk==3.8.1

# Social media APIs
tweepy==4.14.0
google-api-python-client==2.108.0
facebook-sdk==3.1.0

# Utilities
celery==5.3.4
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
loguru==0.7.2
sentry-sdk==1.38.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Async and concurrency
asyncio==3.4.3
aiofiles==23.2.1
aioredis==2.0.1
